+++ src/imio/esign/browser/settings.py
@@ -92,6 +92,13 @@ class IImioEsignSettings(Interface):
             "TAL compliant with variables: view, context, user_data, parapheo_url, request and modules."
         ),
         required=False,
+
+    max_session_size = schema.Int(
+        title=_("Max session size (MB)"),
+        description=_("Maximum size of the session in megabytes. If the total size of files to be signed exceeds this limit, a new session will be created."),
+        default=100,
+        min=1,
+        required=True,
     )