+++ src/imio/esign/browser/table.py
@@ -1,7 +1,9 @@
 # -*- coding: utf-8 -*-
 
 from eea.facetednavigation.interfaces import IFacetedNavigable
+from html import escape
 from imio.esign import _
+from imio.esign.utils import get_state_description
 from plone import api
 from Products.CMFPlone.utils import safe_unicode
 from z3c.table.column import Column
@@ -13,7 +15,8 @@ from zope.i18n import translate
 class IdColumn(Column):
     header = _("ID")
     weight = 10
-    cssClasses = {"th": "th_header_sessions_id"}
+    cssClasses = {"th": "th_header_sessions_id",
+                  "td": "id-column"}
 
     def renderHeadCell(self):
         """
@@ -33,7 +36,7 @@ def external_session_link(session, title=None):
     """Return a tag with the sign external session."""
     title = title or session.get("title", "") or session.get("sign_id", "")
     if not session["sign_id"] or not session["sign_url"]:
-        return u"<p>{0}</p>".format(title)
+        return u"<span>{0}</span>".format(title)
     return u'<a href="{url}" target="_blank">{title}</a>'.format(
         url=session["sign_url"],
         title=safe_unicode(title),
@@ -43,18 +46,22 @@ def external_session_link(session, title=None):
 class StateColumn(Column):
     header = _("State")
     weight = 20
-    cssClasses = {"th": "th_header_sessions_state"}
+    cssClasses = {"th": "th_header_sessions_state",
+                  "td": "state-column"}
 
     def renderCell(self, item):
-        return translate(
-            (item.get("state", "")), context=self.request, default=item.get("state", ""), domain="imio.esign"
-        )
+        state = escape(translate(
+            (item.get("state", "")), context=self.request, default=item.get("state", ""), domain="imio.esign",
+        ))
+        title = escape(translate(get_state_description(item.get("state", "")), context=self.request, domain="imio.esign"))
+        return u"<span title='{title}'>{state}</span>".format(state=state, title=title)
 
 
 class TitleColumn(Column):
     header = _("Title")
     weight = 30
-    cssClasses = {"th": "th_header_sessions_title"}
+    cssClasses = {"th": "th_header_sessions_title",
+                  "td": "title-column"}
 
     def renderCell(self, item):
         title = safe_unicode(item.get("title", ""))
@@ -67,7 +74,8 @@ class TitleColumn(Column):
 class LastUpdateColumn(Column):
     header = _("Last update")
     weight = 40
-    cssClasses = {"th": "th_header_sessions_last_update"}
+    cssClasses = {"th": "th_header_sessions_last_update",
+                  "td": "last-update-column"}
 
     def renderCell(self, item):
         last_update = item.get("last_update")
@@ -78,7 +86,8 @@ class LastUpdateColumn(Column):
 class SignersColumn(Column):
     header = _("Signers")
     weight = 50
-    cssClasses = {"th": "th_header_sessions_signers"}
+    cssClasses = {"th": "th_header_sessions_signers",
+                  "td": "signers-column"}
 
     def renderCell(self, item):
         signers = item.get("signers") or []