+++ CHANGES.rst
@@ -5,8 +5,18 @@ Changelog
 1.0a3 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- Replaced external_session_link p by span.
+  [sgeulette]
+- Switched basic auth to jwt.
+  [chris-adam]
+- Added files size session discriminator
+  [chris-adam, sgeulette]
+- Added title attribute on session state for better description
+  [chris-adam]
+- Raise Unauthorized if accessing @@parapheo and not available
+  [gbastien]
+- Added action to remove a single item from session.
+  [chris-adam]
 
 1.0a2 (2026-02-06)
 ------------------